Refactor into separate cleanup files
